fix bug in fork operation

+++ vm/vm.h
@@ -124,7 +124,7 @@ void						delete_bogie(t_bogie *bogie);
 void						fight(t_champ *champs);
 
 int							get_tdir_big_size(int i);
-short						get_tdir_small_size(int i);
+int							get_tdir_small_size(int i);
 int							get_treg(int scip_size);
 int							get_tind(int size_flag, int scip_size);
 void						move_caret(int steps);
